Change how starter-content is generated.
Starter content is now a DB dump with pre-defined values for the three sites needed for the support environment to work as intended in plugin and theme scenarios as well. This replaces the previous approach to performing a lot of actions with multiple calls to the `wp-env run cli` command, which would be more time-consuming in the long run.